Move individual input event listeners to parent form
Convert the 'controls' div to a form, so we can register an 'input' event
listener on the form and remove the equivallent listeners from each individual
form field. This handles dynamically added form fields too, so no need to
add new event listeners when adding tag groups.

Since the barcode labels are regenerated on all input changes, there's no need
for the template field event listener to clear the pages.