Improve responsive layout for 'large' screens
Large screens where too big for form labels/inputs to span the full width,
but too small to fit the label in a single line of text. Adjust the styling
to always fit the label text in one line.

Also remove the offsets and instead horizontally center the label+input on
larger screens.