Remove the bitCount() method
This method was added when elements where created via a generator,
so that we could compute the bitCount and thus the bitWidth before
generating the elements. With toElements() now returning the full
list at once, we can simply compute the bit count from it.

This is anyways more straight forward if full ASCII code93 is ever
implemented as a character could become one or two symbols worth of
characters.