Only populate a label with content if the inputs are valid
If the inputs (consigner or price) aren't valid then populate an empty label
on the page to ensure the user doesn't try printing invalid labels.

Also update the numberic input parsing to use valueAsNumber and number methods
to ensure it's a positive integer. This is done for both the form validation
CSS class handling and elsewhere. In the process, a bug is fixed that setting
a negative value for a tag's count would subtract from the tags-msg message.

With this change, specifying a non-positive tag price (ex: 0 or empty) inserts
<count> empty tags on the page. This can be useful if a partial page was
previously printed out and now the remaining labels are being used.