Encode the input values in the URL hash
base64 encode a JSON representation of the form inputs and store it
in the URL hash. Then on page load, if present and valid, populate
the form inputs from the URL hash data.

This allows a link to be shared/saved for a given set of inputs
to be reproduced on another machine or at a later date.